perf(#179): W7a — generation→u64 + TaggedPtr index:16|tag:48 (long-run wrap defence)
Two counter-width changes so long-run wraps become unreachable, at zero
hot-path cost. Both are defence-in-depth / future-proofing on the cold
registry protocol, not fixes for a live bug (verified below).

- HeapSlot::generation AtomicU32 → AtomicU64. Written only by the cold
  claim/recycle fetch_add and compared only `== 1` (the first-materialise
  gate); there is NO cached-generation compare anywhere (the stale-TLS/ABA
  hazard is guarded by the TORN sentinel in tls_heap, not by a generation
  compare), so the u32 wrap was never a live ABA vector — but widening removes
  the theoretical 2^32-thread-death wrap for free. REGISTRY_SIZE derives from
  size_of::<Registry>() (slots inline), so the +4 B/slot auto-sizes; slot init
  is name-based (addr_of_mut!), so the layout change is safe. Ir BYTE-IDENTICAL.
- TaggedPtr repack INDEX_BITS 32 → 16 (MAX_HEAPS=4096 needs 13; a const assert
  pins MAX_HEAPS ≤ INDEX_MASK), giving the free_slots ABA tag 48 bits → wrap at
  2^48 ≈ 89 years at 100k pops/s (was the documented probabilistic 2^32). The
  empty sentinel moves 0xFFFF_FFFF → 0xFFFF (still > MAX_HEAPS, no index
  collision; NEXT_FREE_TAIL lives in the per-slot next_free field, never packed
  into the 16-bit half). All TaggedPtr sites are cold (bootstrap/claim/recycle).
  Ir: a uniform −4 (a decrease) from the cheaper bootstrap empty() constant —
  not a hot-path touch; accepted (see docs/perf/IAI_BASELINE.md W7a note).

Boundary-crossing tests (the maintainer's "preset near the limit" ask), new
regression_counter_wrap: tag preset to 2^48−1 → wrap to 0, index round-trips
and never aliases empty; empty sentinel swept vs all boundary indices;
generation preset to u32::MAX−1 → recycle/reclaim across u32::MAX to 2^32,
asserting a value > u32::MAX (only representable in the u64 field — reverting
to u32 wraps to 0 and fails). loom_free_slots_aba updated (INDEX_BITS 16) and
green under --cfg loom. production + all-features suites green, clippy clean.

+++ tests/registry_basic.rs
@@ -87,8 +87,10 @@ fn slot_state(idx: usize) -> u8 {
     reg.slots[idx].state.load(Ordering::Acquire)
 }
 
-/// Read a slot's `generation` atomically (test helper).
-fn slot_generation(idx: usize) -> u32 {
+/// Read a slot's `generation` atomically (test helper). `generation` is
+/// `AtomicU64` since task W7a (widened from `AtomicU32` to move the recycle→
+/// reclaim ABA wrap from `2^32` to an unreachable `2^64`).
+fn slot_generation(idx: usize) -> u64 {
     let reg = bootstrap::ensure();
     reg.slots[idx].generation.load(Ordering::Acquire)
 }
